refactor: extract options account UI
